Split the one larger file int smaller files and separate concerns better.
