+++ src/wrp-c.c
@@ -6,12 +6,13 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <msgpack.h>
-#include <trower-base64/base64.h>
 #include <cimplog/cimplog.h>
 
-#include "wrp-c.h"
+#include "b64.h"
+#include "msgpack.h"
+#include "string.h"
 #include "utils.h"
+#include "wrp-c.h"
 
 /*----------------------------------------------------------------------------*/
 /*                                   Macros                                   */
@@ -25,99 +26,16 @@
 /*----------------------------------------------------------------------------*/
 /*                               Data Structures                              */
 /*----------------------------------------------------------------------------*/
-struct wrp_token {
-    const char *name;
-    size_t length;
-};
-
-
-struct req_res_t {
-    int msgType;
-    int statusValue;
-    char* source;
-    char* dest;
-    char* transaction_uuid;
-    partners_t *partner_ids;
-    headers_t *headers;
-    void *payload;
-    size_t payload_size;
-    bool include_spans;
-    struct money_trace_spans spans;
-    data_t *metadata;
-    int status;
-    int rdr;
-    char *path;
-    char *service_name;
-    char *url;
-    char *content_type;
-    char *accept;
-};
-
 
 /*----------------------------------------------------------------------------*/
 /*                            File Scoped Variables                           */
 /*----------------------------------------------------------------------------*/
-static const char *__empty_list = "''";
-#define METADATA_MAP_SIZE                          1
-/* Reminder: sizeof("string") includes the trailing '\0', which we don't want. */
-static const struct wrp_token WRP_MSG_TYPE      = { .name = "msg_type", .length = sizeof( "msg_type" ) - 1 };
-static const struct wrp_token WRP_STATUS        = { .name = "status", .length = sizeof( "status" ) - 1 };
-static const struct wrp_token WRP_SOURCE        = { .name = "source", .length = sizeof( "source" ) - 1 };
-static const struct wrp_token WRP_DEST          = { .name = "dest", .length = sizeof( "dest" ) - 1 };
-static const struct wrp_token WRP_TRANS_ID      = { .name = "transaction_uuid", .length = sizeof( "transaction_uuid" ) - 1 };
-static const struct wrp_token WRP_CONTENT_TYPE  = { .name = "content_type", .length = sizeof( "content_type" ) - 1 };
-static const struct wrp_token WRP_ACCEPT        = { .name = "accept", .length = sizeof( "accept" ) - 1 };
-static const struct wrp_token WRP_HEADERS       = { .name = "headers", .length = sizeof( "headers" ) - 1 };
-static const struct wrp_token WRP_PAYLOAD       = { .name = "payload", .length = sizeof( "payload" ) - 1 };
-static const struct wrp_token WRP_SPANS         = { .name = "spans", .length = sizeof( "spans" ) - 1 };
-static const struct wrp_token WRP_INCLUDE_SPANS = { .name = "include_spans", .length = sizeof( "include_spans" ) - 1 };
-static const struct wrp_token WRP_SERVICE_NAME  = { .name = "service_name", .length = sizeof( "service_name" ) - 1 };
-static const struct wrp_token WRP_URL           = { .name = "url", .length = sizeof( "url" ) - 1 };
-static const struct wrp_token WRP_METADATA      = { .name = "metadata", .length = sizeof( "metadata" ) - 1 };
-static const struct wrp_token WRP_RDR           = { .name = "rdr", .length = sizeof( "rdr" ) - 1 };
-static const struct wrp_token WRP_PATH          = { .name = "path", .length = sizeof( "path" ) - 1 };
-static const struct wrp_token WRP_PARTNER_IDS   = { .name = "partner_ids", .length = sizeof( "partner_ids" ) - 1 };
-static const int WRP_MAP_SIZE                   = 4; // mandatory msg_type,source,dest,payload
+#define METADATA_MAP_SIZE   1
 
 
 /*----------------------------------------------------------------------------*/
 /*                             Function Prototypes                            */
 /*----------------------------------------------------------------------------*/
-static ssize_t __wrp_struct_to_bytes( const wrp_msg_t *msg, char **bytes );
-static ssize_t __wrp_struct_to_base64( const wrp_msg_t *msg, char **bytes );
-static ssize_t __wrp_struct_to_string( const wrp_msg_t *msg, char **bytes );
-static ssize_t __wrp_keep_alive_to_string( char **bytes );
-static ssize_t __wrp_auth_struct_to_string( const struct wrp_auth_msg *auth,
-        char **bytes );
-static ssize_t __wrp_req_struct_to_string( const struct wrp_req_msg *req, char **bytes );
-static ssize_t __wrp_event_struct_to_string( const struct wrp_event_msg *event,
-        char **bytes );
-static char* __get_header_string( headers_t *headers );
-static char* __get_spans_string( const struct money_trace_spans *spans );
-static char* __get_partner_ids_string( partners_t *partner_ids );
-static void __special_free( char *s );
-static void __msgpack_pack_string_nvp( msgpack_packer *pk,
-                                       const struct wrp_token *token,
-                                       const char *val );
-static void __msgpack_pack_string( msgpack_packer *pk, const void *string, size_t n );
-
-static ssize_t __wrp_pack_structure( struct req_res_t *encodeReq, char **data );
-
-
-static ssize_t __wrp_base64_to_struct( const void *base64_data, const size_t base64_size,
-                                       wrp_msg_t **msg_ptr );
-static ssize_t __wrp_bytes_to_struct( const void *bytes, const size_t length,
-                                      wrp_msg_t **msg_ptr );
-
-static void decodeRequest( msgpack_object deserialized, struct req_res_t **decodeReq );
-static char* getKey_MsgtypeStr( const msgpack_object key, const size_t keySize,
-                                char* keyString );
-static char* getKey_MsgtypeBin( const msgpack_object key, const size_t binSize,
-                                char* keyBin );
-static void __msgpack_maps( msgpack_packer *pk, const data_t *dataMap );
-static void decodeMapRequest( msgpack_object deserialized,
-                              struct req_res_t **decodeMapReq );
-static void mapCommonString( msgpack_packer *pk, struct req_res_t *encodeComReq );
 static int alterMap( char * buf );
 
 
@@ -137,15 +55,15 @@ ssize_t wrp_struct_to( const wrp_msg_t *msg, const enum wrp_format fmt, void **b
 
     switch( fmt ) {
         case WRP_BYTES:
-            rv = __wrp_struct_to_bytes( msg, &data );
+            rv = wrp_struct_to_bytes( msg, &data );
             break;
 
         case WRP_BASE64:
-            rv = __wrp_struct_to_base64( msg, &data );
+            rv = wrp_struct_to_base64( msg, &data );
             break;
 
         case WRP_STRING:
-            rv = __wrp_struct_to_string( msg, &data );
+            rv = internal_struct_to_string( msg, &data );
             break;
 
         default:
@@ -164,7 +82,7 @@ char* wrp_struct_to_string( const wrp_msg_t *msg )
     char *string = NULL;
 
     if( NULL != msg ) {
-        __wrp_struct_to_string( msg, &string );
+        internal_struct_to_string( msg, &string );
     }
 
     return string;
@@ -182,11 +100,11 @@ ssize_t wrp_to_struct( const void *bytes, const size_t length,
 
     switch( fmt ) {
         case WRP_BYTES:
-            rv = __wrp_bytes_to_struct( bytes, length, msg );
+            rv = wrp_bytes_to_struct( bytes, length, msg );
             break;
 
         case WRP_BASE64:
-            rv = __wrp_base64_to_struct( bytes, length, msg );
+            rv = wrp_base64_to_struct( bytes, length, msg );
             break;
 
         default:
@@ -448,20 +366,20 @@ const char *wrp_get_msg_source( const wrp_msg_t *wrp_msg )
 
 /* See wrp-c.h for details. */
 char *wrp_get_msg_element( const enum wrp_device_id_element element,
-                           const wrp_msg_t *wrp_msg, const enum wrp_token_name wrp_token )
+                           const wrp_msg_t *wrp_msg, const enum wrp_token_name token )
 {
     const char *dest;
     const char *source;
     const char *start = NULL, *end = NULL;
     char *rv = NULL;
 
-    if( wrp_token == DEST ) {
+    if( token == DEST ) {
         dest = wrp_get_msg_dest( wrp_msg );
 
         if( NULL != dest ) {
             start = dest;
         }
-    } else if( wrp_token == SOURCE ) {
+    } else if( token == SOURCE ) {
         source = wrp_get_msg_source( wrp_msg );
 
         if( NULL != source ) {
@@ -554,1428 +472,10 @@ int wrp_does_service_match( const char *service, const char *device_id )
     return -1;
 }
 
+
 /*----------------------------------------------------------------------------*/
 /*                             Internal functions                             */
 /*----------------------------------------------------------------------------*/
-/**
- *  Do the work of converting into msgpack binary format.
- *
- *  @param msg   [in]  the message to convert
- *  @param bytes [out] the place to put the output (never NULL)
- *
- *  @return size of buffer (in bytes) on success or less then 1 on error
- */
-
-static ssize_t __wrp_struct_to_bytes( const wrp_msg_t *msg, char **bytes )
-{
-    ssize_t rv = -1;
-    const struct wrp_req_msg *req = & ( msg->u.req );
-    const struct wrp_event_msg *event = & ( msg->u.event );
-    const struct wrp_svc_registration_msg *reg = & ( msg->u.reg );
-    const struct wrp_crud_msg *crud = &( msg->u.crud );
-    const struct wrp_auth_msg *auth = &( msg->u.auth );
-
-    if( NULL == msg || NULL == bytes ) {
-        return -1;
-    }
-
-    *bytes = NULL;
-    struct req_res_t *encode = calloc( 1, sizeof(struct req_res_t) );
-
-    if( encode != NULL ) {
-
-        //convert to binary bytes using msgpack
-        switch( msg->msg_type ) {
-            case WRP_MSG_TYPE__AUTH:
-                encode->msgType = msg->msg_type;
-                encode->statusValue = auth->status;
-                rv = __wrp_pack_structure( encode, bytes );
-                break;
-
-            case WRP_MSG_TYPE__REQ:
-                encode->source = req->source;
-                encode->dest = req->dest;
-                encode->content_type = req->content_type;
-                encode->accept = req->accept;
-                encode->transaction_uuid = req->transaction_uuid;
-                encode->include_spans = req->include_spans;
-                encode->spans = req->spans;
-                encode->payload = req->payload;//void
-                encode->payload_size = req->payload_size;
-                encode->headers = req->headers;
-                encode->metadata = req->metadata;
-                encode->partner_ids = req->partner_ids;
-                encode->msgType = msg->msg_type;
-                rv = __wrp_pack_structure( encode, bytes );
-                break;
-
-            case WRP_MSG_TYPE__EVENT:
-                encode->source = event->source;
-                encode->dest = event->dest;
-                encode->content_type = event->content_type;
-                encode->transaction_uuid = NULL;
-                encode->include_spans = false;
-                encode->spans.spans = NULL;
-                encode->payload = event->payload;//void
-                encode->payload_size = event->payload_size;
-                encode->headers = event->headers;
-                encode->metadata = event->metadata;
-                encode->partner_ids = event->partner_ids;
-                encode->msgType = msg->msg_type;
-                rv = __wrp_pack_structure( encode, bytes );
-                break;
-
-            case WRP_MSG_TYPE__SVC_REGISTRATION:
-                encode->msgType = msg->msg_type;
-                encode->service_name = reg->service_name;
-                encode->url = reg->url;
-                encode->transaction_uuid = NULL;
-                encode->include_spans = false;
-                encode->spans.spans = NULL;
-                rv = __wrp_pack_structure( encode, bytes );
-                break;
-
-            case WRP_MSG_TYPE__CREATE:
-            case WRP_MSG_TYPE__RETREIVE:
-            case WRP_MSG_TYPE__UPDATE:
-            case WRP_MSG_TYPE__DELETE:
-                encode->msgType = msg->msg_type;
-                encode->source = crud->source;
-                encode->dest = crud->dest;
-                encode->transaction_uuid = crud->transaction_uuid;
-                encode->include_spans = crud->include_spans;
-                encode->spans = crud->spans;
-                encode->content_type = crud->content_type;
-                encode->accept = crud->accept;
-                encode->payload = crud->payload;//void
-                encode->payload_size = crud->payload_size;
-                encode->partner_ids = crud->partner_ids;
-                encode->headers = crud->headers;
-                encode->metadata = crud->metadata;
-                encode->path = crud->path;
-                encode->status = crud->status;
-                encode->rdr = crud->rdr;
-                rv = __wrp_pack_structure( encode, bytes );
-                break;
-
-            case WRP_MSG_TYPE__SVC_ALIVE:
-                encode->msgType = msg->msg_type;
-                rv = __wrp_pack_structure( encode, bytes );
-                break;
-
-            default:
-                WRP_ERROR( "Unknown msgType to encode\n" );
-                break;
-        }
-
-        free( encode );
-        return rv;
-    } else {
-        WRP_ERROR( "Memory allocation failed\n" );
-    }
-
-    return -1;
-}
-
-
-/**
- *  Do the work of converting into base64 format.
- *
- *  @param msg   [in]  the message to convert
- *  @param bytes [out] the place to put the output (never NULL)
- *
- *  @return size of buffer (in bytes) on success or less then 1 on error
- */
-static ssize_t __wrp_struct_to_base64( const wrp_msg_t *msg, char **bytes )
-{
-    ssize_t data_len;
-    ssize_t rv = -1;
-    char *data = NULL;
-
-    data_len = __wrp_struct_to_bytes( msg, &data );
-    if( 0 < data_len ) {
-        size_t b64_len = 0;
-        char *base64 = NULL;
-
-        base64 = b64_encode_with_alloc( (uint8_t*) data, data_len, &b64_len );
-
-        if( base64 ) {
-            *bytes = base64;
-            rv = b64_len;
-        }
-    }
-
-    if( data ) {
-        free( data );
-    }
-
-    return rv;
-}
-
-
-/**
- *  Split out the different types to different functions.
- *
- *  @param msg   [in]  the message to convert
- *  @param bytes [out] the output of the conversion
- *
- *  @return the number of bytes in the string or less then 1 on error
- */
-static ssize_t __wrp_struct_to_string( const wrp_msg_t *msg, char **bytes )
-{
-    switch( msg->msg_type ) {
-        case WRP_MSG_TYPE__SVC_ALIVE:
-            return __wrp_keep_alive_to_string( bytes );
-
-        case WRP_MSG_TYPE__AUTH:
-            return __wrp_auth_struct_to_string( &msg->u.auth, bytes );
-
-        case WRP_MSG_TYPE__REQ:
-            return __wrp_req_struct_to_string( &msg->u.req, bytes );
-
-        case WRP_MSG_TYPE__EVENT:
-            return __wrp_event_struct_to_string( &msg->u.event, bytes );
-
-        default:
-            break;
-    }
-
-    return -1;
-}
-
-static ssize_t __wrp_keep_alive_to_string( char **bytes )
-{
-    const char *keep_alive_fmt = "wrp_keep_alive_msg {\n"
-                                 "}\n";
-    size_t len;
-
-    *bytes = mlaprintf( &len, keep_alive_fmt );
-    if( NULL == *bytes ) {
-        len = -1;
-    }
-
-    return len;
-}
-
-/**
- *  Convert the auth structure to a string.
- *
- *  @param auth  [in]  the message to convert
- *  @param bytes [out] the output of the conversion
- *
- *  @return the number of bytes in the string or less then 1 on error
- */
-static ssize_t __wrp_auth_struct_to_string( const struct wrp_auth_msg *auth,
-        char **bytes )
-{
-    const char *auth_fmt = "wrp_auth_msg {\n"
-                           "    .status = %d\n"
-                           "}\n";
-    size_t len;
-
-    *bytes = mlaprintf( &len, auth_fmt, auth->status );
-    if( NULL == *bytes ) {
-        len = -1;
-    }
-
-    return len;
-}
-
-
-/**
- *  Convert the req structure to a string.
- *
- *  @param msg   [in]  the message to convert
- *  @param bytes [out] the output of the conversion
- *
- *  @return the number of bytes in the string or less then 1 on error
- */
-static ssize_t __wrp_req_struct_to_string( const struct wrp_req_msg *req, char **bytes )
-{
-    const char *req_fmt = "wrp_req_msg {\n"
-                          "    .transaction_uuid = %s\n"
-                          "    .source           = %s\n"
-                          "    .dest             = %s\n"
-                          "    .partner_ids      = %s\n"
-                          "    .headers          = %s\n"
-                          "    .content_type     = %s\n"
-                          "    .accept           = %s\n"
-                          "    .include_spans    = %s\n"
-                          "    .spans            = %s\n"
-                          "    .payload_size     = %zd\n"
-                          "}\n";
-    size_t len;
-    char *headers;
-    char *spans;
-    char *partner_ids;
-
-    headers = __get_header_string( req->headers );
-    spans = __get_spans_string( &req->spans );
-    partner_ids = __get_partner_ids_string( req->partner_ids );
-
-    *bytes = mlaprintf( &len, req_fmt, req->transaction_uuid, req->source,
-                        req->dest, partner_ids, headers, req->content_type,
-                        req->accept, ( req->include_spans ? "true" : "false" ),
-                        spans, req->payload_size );
-    if( NULL == *bytes ) {
-        len = -1;
-    }
-
-    __special_free( headers );
-    __special_free( spans );
-    __special_free( partner_ids );
-
-    return len;
-}
-
-
-/**
- *  Convert the event structure to a string.
- *
- *  @param event [in]  the message to convert
- *  @param bytes [out] the output of the conversion
- *
- *  @return the number of bytes in the string or less then 1 on error
- */
-static ssize_t __wrp_event_struct_to_string( const struct wrp_event_msg *event,
-        char **bytes )
-{
-    const char *event_fmt = "wrp_event_msg {\n"
-                            "    .source           = %s\n"
-                            "    .dest             = %s\n"
-                            "    .partner_ids      = %s\n"
-                            "    .headers          = %s\n"
-                            "    .content_type     = %s\n"
-                            "    .payload_size     = %zd\n"
-                            "}\n";
-    size_t len;
-    char *headers;
-    char *partner_ids;
-
-    headers = __get_header_string( event->headers );
-    partner_ids = __get_partner_ids_string( event->partner_ids );
-
-    *bytes = mlaprintf( &len, event_fmt, event->source, event->dest, partner_ids,
-                        headers, event->content_type, event->payload_size );
-    if( NULL == *bytes ) {
-        len = -1;
-    }
-
-    __special_free( headers );
-    __special_free( partner_ids );
-
-    return len;
-}
-
-
-/**
- *  Converts the list of headers into a string to print.
- *
- *  @note The caller must check to see if the value is equal to __empty_list
- *        and not free it if equal.  If not equal it must be freed.
- *  @note This function never returns NULL.
- *
- *  @param headers [in] the headers to make into a string
- *
- *  @return The string representation of the headers.
- */
-static char* __get_header_string( headers_t *headers )
-{
-    char *rv;
-    rv = ( char* ) __empty_list;
-
-    if( headers ) {
-        char *tmp;
-        size_t i;
-        int comma;
-        size_t length;
-        comma = 0;
-        length = 2; /* For ' characters. */
-
-        for( i = 0; i < headers->count; i++ ) {
-            length += comma;
-            length += strlen( headers->headers[i] );
-            comma = 2;
-        }
-
-        tmp = malloc( sizeof(char) * (length + 1) );   /* +1 for '\0' */
-
-        if( NULL != tmp ) {
-            const char *comma;
-            rv = tmp;
-            comma = "";
-            *tmp = '\0';
-            tmp = strcat( tmp, "'" );
-
-            for( i = 0; i < headers->count; i++ ) {
-                tmp = strcat( tmp, comma );
-                tmp = strcat( tmp, headers->headers[i] );
-                comma = ", ";
-            }
-
-            tmp = strcat( tmp, "'" );
-        }
-    }
-
-    return rv;
-}
-
-static char* __get_partner_ids_string( partners_t *partner_ids )
-{
-    char *rv;
-    rv = ( char* ) __empty_list;
-
-    if( partner_ids ) {
-        char *tmp;
-        size_t i;
-        int comma;
-        size_t length;
-        comma = 0;
-        length = 2; /* For ' characters. */
-
-        for( i = 0; i < partner_ids->count; i++ ) {
-            length += comma;
-            length += strlen( partner_ids->partner_ids[i] );
-            comma = 2;
-        }
-
-        tmp = malloc( sizeof(char) * (length + 1) );   /* +1 for '\0' */
-
-        if( NULL != tmp ) {
-            const char *comma;
-            rv = tmp;
-            comma = "";
-            *tmp = '\0';
-            tmp = strcat( tmp, "'" );
-
-            for( i = 0; i < partner_ids->count; i++ ) {
-                tmp = strcat( tmp, comma );
-                tmp = strcat( tmp, partner_ids->partner_ids[i] );
-                comma = ", ";
-            }
-
-            tmp = strcat( tmp, "'" );
-        }
-    }
-
-    return rv;
-}
-
-static void __special_free( char *s )
-{
-    if( __empty_list != s ) {
-        free( s );
-    }
-}
-
-static void __msgpack_headers( msgpack_packer *pk, headers_t *headers )
-{
-    if( NULL != headers ) {
-        size_t count = headers->count;
-        __msgpack_pack_string( pk, WRP_HEADERS.name, WRP_HEADERS.length );
-        msgpack_pack_array( pk, count );
-        count = 0;
-
-        while( count < headers->count ) {
-            __msgpack_pack_string( pk, headers->headers[count], strlen( headers->headers[count] ) );
-            count++;
-        }
-    }
-}
-
-static void __msgpack_partner_ids( msgpack_packer *pk, partners_t *partner_ids )
-{
-    if( NULL != partner_ids ) {
-        size_t count = partner_ids->count;
-        __msgpack_pack_string( pk, WRP_PARTNER_IDS.name, WRP_PARTNER_IDS.length );
-        msgpack_pack_array( pk, count );
-        count = 0;
-
-        while( count < partner_ids->count ) {
-            __msgpack_pack_string( pk, partner_ids->partner_ids[count],
-                                   strlen( partner_ids->partner_ids[count] ) );
-            count++;
-        }
-    }
-}
-
-static void __msgpack_spans( msgpack_packer *pk, const struct money_trace_spans *spans )
-{
-    if( ( NULL != spans ) && ( 0 < spans->count ) ) {
-        size_t i;
-        struct money_trace_span *span;
-        __msgpack_pack_string( pk, WRP_SPANS.name, WRP_SPANS.length );
-        msgpack_pack_array( pk, spans->count );
-        span = spans->spans;
-
-        for( i = 0; i < spans->count; i++ ) {
-            msgpack_pack_array( pk, 3 );
-            __msgpack_pack_string( pk, span->name, strlen( span->name ) );
-            msgpack_pack_uint64( pk, span->start );
-            msgpack_pack_uint32( pk, span->duration );
-            span++;
-        }
-    }
-}
-
-static void __msgpack_maps( msgpack_packer *pk, const data_t *dataMap )
-{
-    size_t i;
-
-    if( dataMap != NULL ) {
-        struct data *tmpdata;
-        msgpack_pack_map( pk, dataMap->count );
-        tmpdata = dataMap->data_items;
-        WRP_DEBUG( "dataMap->count is %zu\n", dataMap->count );
-
-        for( i = 0; i < dataMap->count; i++ ) {
-            struct wrp_token WRP_MAP_NAME;
-
-            WRP_MAP_NAME.name = tmpdata[i].name;
-            WRP_MAP_NAME.length = strlen( tmpdata[i].name );
-            __msgpack_pack_string_nvp( pk, &WRP_MAP_NAME, tmpdata[i].value );
-        }
-    } else {
-        WRP_ERROR( "Map is NULL.Do not pack\n" );
-    }
-}
-
-static void __msgpack_pack_string_nvp( msgpack_packer *pk,
-                                       const struct wrp_token *token,
-                                       const char *val )
-{
-    if( ( NULL != token ) && ( NULL != val ) ) {
-        __msgpack_pack_string( pk, token->name, token->length );
-        __msgpack_pack_string( pk, val, strlen( val ) );
-    }
-}
-
-static void __msgpack_pack_string( msgpack_packer *pk, const void *string, size_t n )
-{
-    msgpack_pack_str( pk, n );
-    msgpack_pack_str_body( pk, string, n );
-}
-
-//Pack msgType,source,dest,headers,metadata,partner_ids
-static void mapCommonString( msgpack_packer *pk, struct req_res_t *encodeComReq )
-{
-    __msgpack_pack_string( pk, WRP_MSG_TYPE.name, WRP_MSG_TYPE.length );
-    msgpack_pack_int( pk, encodeComReq->msgType );
-    __msgpack_pack_string_nvp( pk, &WRP_SOURCE, encodeComReq->source );
-    __msgpack_pack_string_nvp( pk, &WRP_DEST, encodeComReq->dest );
-    __msgpack_partner_ids( pk, encodeComReq->partner_ids );
-    __msgpack_headers( pk, encodeComReq->headers );
-
-    if( encodeComReq->metadata != NULL ) {
-        __msgpack_pack_string( pk, WRP_METADATA.name, WRP_METADATA.length );
-        __msgpack_maps( pk, encodeComReq->metadata );
-    }
-}
-
-/**
- *  Encode/Pack the wrp message structure using msgpack
- *
- *  @param  [in]  the messages to convert
- *  @param data [out] the encoded output
- *
- *  @return the number of bytes in the string or less then 1 on error*/
-
-
-static ssize_t __wrp_pack_structure( struct req_res_t *encodeReq, char **data )
-
-{
-    msgpack_sbuffer sbuf;
-    msgpack_packer pk;
-    ssize_t rv;
-    int wrp_map_size = WRP_MAP_SIZE;
-    struct req_res_t *encodeReqtmp =  encodeReq;
-    /***   Start of Msgpack Encoding  ***/
-    WRP_DEBUG( "***   Start of Msgpack Encoding  ***\n" );
-    msgpack_sbuffer_init( &sbuf );
-    msgpack_packer_init( &pk, &sbuf, msgpack_sbuffer_write );
-
-    // Change wrp_map_size value depending on if optional fields spans and headers,metadata,content_type,accept crud payload are present
-    if( encodeReqtmp->transaction_uuid ) {
-        wrp_map_size++;
-    }
-
-    if( encodeReqtmp->content_type ) {
-        wrp_map_size++;
-    }
-
-    if( encodeReqtmp->accept ) {
-        wrp_map_size++;
-    }
-
-    if( encodeReqtmp->include_spans ) {
-        wrp_map_size++;
-    }
-
-    if( ( NULL != encodeReqtmp->spans.spans ) && ( 0 < encodeReqtmp->spans.count ) ) {
-        wrp_map_size++;
-    }
-
-    if( encodeReqtmp->headers ) {
-        wrp_map_size++;
-    }
-
-    if( encodeReqtmp->metadata != NULL ) {
-        wrp_map_size++;
-    }
-
-    if( encodeReqtmp->partner_ids ) {
-        wrp_map_size++;
-    }
-
-    switch( encodeReqtmp->msgType ) {
-        case WRP_MSG_TYPE__AUTH:
-            wrp_map_size = 2;//Hardcoded. Pack for msgType and status alone for auth msg
-            msgpack_pack_map( &pk, wrp_map_size );
-            __msgpack_pack_string( &pk, WRP_MSG_TYPE.name, WRP_MSG_TYPE.length );
-            msgpack_pack_int( &pk, encodeReqtmp->msgType );
-            __msgpack_pack_string( &pk, WRP_STATUS.name, WRP_STATUS.length );
-            msgpack_pack_int( &pk, encodeReqtmp->statusValue );
-            break;
-
-        case WRP_MSG_TYPE__REQ:
-            msgpack_pack_map( &pk, wrp_map_size );
-            //Pack msgType,source,dest,headers,metadata,partner_ids
-            mapCommonString( &pk, encodeReqtmp );
-            __msgpack_pack_string_nvp( &pk, &WRP_TRANS_ID, encodeReqtmp->transaction_uuid );
-            __msgpack_pack_string_nvp( &pk, &WRP_CONTENT_TYPE, encodeReqtmp->content_type );
-
-            if( encodeReqtmp->accept ) {
-                __msgpack_pack_string_nvp( &pk, &WRP_ACCEPT, encodeReqtmp->accept );
-            }
-
-            if( encodeReqtmp->include_spans ) {
-                __msgpack_pack_string( &pk, WRP_INCLUDE_SPANS.name, WRP_INCLUDE_SPANS.length );
-                msgpack_pack_true( &pk );
-            }
-
-            __msgpack_spans( &pk, &encodeReqtmp->spans );
-            __msgpack_pack_string( &pk, WRP_PAYLOAD.name, WRP_PAYLOAD.length );
-            msgpack_pack_bin( &pk, encodeReqtmp->payload_size );
-            msgpack_pack_bin_body( &pk, encodeReqtmp->payload, encodeReqtmp->payload_size );
-            break;
-
-        case WRP_MSG_TYPE__EVENT:
-            msgpack_pack_map( &pk, wrp_map_size );
-            //Pack msgType,source,dest,headers,metadata,partner_ids
-            mapCommonString( &pk, encodeReqtmp );
-            __msgpack_pack_string_nvp( &pk, &WRP_CONTENT_TYPE, encodeReqtmp->content_type );
-            __msgpack_pack_string( &pk, WRP_PAYLOAD.name, WRP_PAYLOAD.length );
-            msgpack_pack_bin( &pk, encodeReqtmp->payload_size );
-            msgpack_pack_bin_body( &pk, encodeReqtmp->payload, encodeReqtmp->payload_size );
-            break;
-
-        case WRP_MSG_TYPE__SVC_REGISTRATION:
-            wrp_map_size = 3;//Hardcoded.Pack service name and url only
-            msgpack_pack_map( &pk, wrp_map_size );
-            __msgpack_pack_string( &pk, WRP_MSG_TYPE.name, WRP_MSG_TYPE.length );
-            msgpack_pack_int( &pk, encodeReqtmp->msgType );
-            __msgpack_pack_string_nvp( &pk, &WRP_SERVICE_NAME, encodeReqtmp->service_name );
-            __msgpack_pack_string_nvp( &pk, &WRP_URL, encodeReqtmp->url );
-            break;
-
-        case WRP_MSG_TYPE__SVC_ALIVE:
-            wrp_map_size = 1;//Hardcoded. Pack for msgType only
-            msgpack_pack_map( &pk, wrp_map_size );
-            __msgpack_pack_string( &pk, WRP_MSG_TYPE.name, WRP_MSG_TYPE.length );
-            msgpack_pack_int( &pk, encodeReqtmp->msgType );
-            break;
-
-        case WRP_MSG_TYPE__CREATE:
-        case WRP_MSG_TYPE__RETREIVE:
-        case WRP_MSG_TYPE__UPDATE:
-        case WRP_MSG_TYPE__DELETE:
-
-            if( encodeReqtmp->payload == NULL ) {
-                wrp_map_size--;
-                WRP_INFO( "CRUD payload is NULL map size is %d\n", wrp_map_size );
-            }
-
-            if( encodeReqtmp->path != NULL ) {
-                wrp_map_size++;
-            }
-
-            if( encodeReqtmp->status != 0 ) {
-                wrp_map_size++; //status
-            }
-
-            if( encodeReqtmp->rdr >= 0 ) {
-                wrp_map_size++; //rdr
-            }
-
-            msgpack_pack_map( &pk, wrp_map_size );
-            //Pack msgType,source,dest,headers,metadata,partner_ids
-            mapCommonString( &pk, encodeReqtmp );
-            __msgpack_pack_string_nvp( &pk, &WRP_TRANS_ID, encodeReqtmp->transaction_uuid );
-
-            if( encodeReqtmp->include_spans ) {
-                __msgpack_pack_string( &pk, WRP_INCLUDE_SPANS.name, WRP_INCLUDE_SPANS.length );
-                msgpack_pack_true( &pk );
-            }
-
-            __msgpack_spans( &pk, &encodeReqtmp->spans );
-
-            if( encodeReqtmp->status != 0 ) {
-                __msgpack_pack_string( &pk, WRP_STATUS.name, WRP_STATUS.length );
-                msgpack_pack_int( &pk, encodeReqtmp->status );
-            }
-
-            if( encodeReqtmp->rdr >= 0 ) {
-                __msgpack_pack_string( &pk, WRP_RDR.name, WRP_RDR.length );
-                msgpack_pack_int( &pk, encodeReqtmp->rdr );
-            }
-
-            if( encodeReqtmp->path != NULL ) {
-                __msgpack_pack_string_nvp( &pk, &WRP_PATH, encodeReqtmp->path );
-            }
-
-            if( encodeReqtmp->content_type != NULL ) {
-                __msgpack_pack_string_nvp( &pk, &WRP_CONTENT_TYPE, encodeReqtmp->content_type );
-            }
-
-            if( encodeReqtmp->accept != NULL ) {
-                __msgpack_pack_string_nvp( &pk, &WRP_ACCEPT, encodeReqtmp->accept );
-            }
-
-            if( encodeReqtmp->payload != NULL ) {
-                __msgpack_pack_string( &pk, WRP_PAYLOAD.name, WRP_PAYLOAD.length );
-                msgpack_pack_bin( &pk, encodeReqtmp->payload_size );
-                msgpack_pack_bin_body( &pk, encodeReqtmp->payload, encodeReqtmp->payload_size );
-            }
-
-            break;
-
-        default:
-            WRP_ERROR( "Un-supported format to pack\n" );
-            return -1;
-    }
-
-    rv = -1;
-
-    if( sbuf.data ) {
-        *data = malloc( sizeof(char) * sbuf.size );
-
-        if( NULL != *data ) {
-            memcpy( *data, sbuf.data, sbuf.size );
-            rv = sbuf.size;
-        }
-    }
-
-    msgpack_sbuffer_destroy( &sbuf );
-    return rv;
-}
-
-/**
- *  Encode/pack only metadata from wrp_msg_t structure.
- *
- *  @note Do not call free of output data in failure case!
- *
- *  @param msg [in] packData the data_t structure to pack/encode
- *  @param msg [out] the encoded output
- *  @return encoded buffer size or less than 1 in failure case
- */
-
-ssize_t wrp_pack_metadata( const data_t *packData, void **data )
-{
-    size_t rv = -1;
-    msgpack_sbuffer sbuf;
-    msgpack_packer pk;
-    msgpack_sbuffer_init( &sbuf );
-    msgpack_packer_init( &pk, &sbuf, msgpack_sbuffer_write );
-
-    if( packData != NULL && packData->count != 0 ) {
-        __msgpack_pack_string( &pk, WRP_METADATA.name, WRP_METADATA.length );
-        __msgpack_maps( &pk, packData );
-    } else {
-        WRP_ERROR( "Metadata is NULL\n" );
-        return rv;
-    }
-
-    if( sbuf.data ) {
-        *data = malloc( sizeof(char) * sbuf.size );
-
-        if( NULL != *data ) {
-            memcpy( *data, sbuf.data, sbuf.size );
-            rv = sbuf.size;
-        }
-    }
-
-    msgpack_sbuffer_destroy( &sbuf );
-    return rv;
-}
-
-
-/**
- *  Converts the list of times into a string to print.
- *
- *  @note The caller must check to see if the value is equal to __empty_list
- *        and not free it if equal.  If not equal it must be freed.
- *  @note This function never returns NULL.
- *
- *  @param spans [in] the spans to make into a string
- *
- *  @return The string representation of the times.
- */
-static char* __get_spans_string( const struct money_trace_spans *spans )
-{
-    char *rv;
-    size_t count;
-    count = spans->count;
-    rv = ( char* ) __empty_list;
-
-    if( 0 < count ) {
-        char *buffer;
-        size_t length, i;
-        const struct money_trace_span *p;
-        length = 0;
-        p = spans->spans;
-
-        for( i = 0; i < count; i++, p++ ) {
-            length += snprintf( NULL, 0, "\n        %s: %" PRIu64 " - %" PRIu32,
-                                p->name, p->start, p->duration );
-        }
-
-        buffer = malloc( sizeof(char) * ( length + 1 ) );   /* +1 for '\0' */
-
-        if( NULL != buffer ) {
-            rv = buffer;
-            p = spans->spans;
-
-            for( i = 0; i < count; i++, p++ ) {
-                buffer += sprintf( buffer, "\n        %s: %" PRIu64 " - %" PRIu32,
-                                   p->name, p->start, p->duration );
-            }
-
-            *buffer = '\0';
-        }
-    }
-
-    return rv;
-}
-
-/**
- * @brief decodeRequest function to unpack the request received from server.
- *
- * @param[in] deserialized msgpack object to deserialize
- * @param[in] msg_type type of the message
- * @param[in] source_ptr source from which the response is generated
- * @param[in] dest_ptr Destination to which it has to be sent
- * @param[in] transaction_uuid_ptr Id
- * @param[in] statusValue status value of authorization request
- * @param[in] payload_ptr bin payload to be extracted from request
- * @param[in] include_spans to include timing values
- */
-
-static void decodeRequest( msgpack_object deserialized, struct req_res_t **decodeReq )
-{
-    unsigned int i = 0;
-    int keySize = 0;
-    char* keyString = NULL;
-    char* keyName = NULL;
-    int binValueSize = 0, StringValueSize = 0;
-    char* NewStringVal, *StringValue;
-    char* keyValue = NULL;
-    char *transaction_uuid = NULL;
-    char *source = NULL;
-    char *dest = NULL;
-    char *payload = NULL;
-    char *service_name = NULL;
-    char *url = NULL;
-    char *path = NULL;
-    char *content_type = NULL;
-    char *headers = NULL;
-    char *partnerID = NULL;
-    char *accept = NULL;
-    struct req_res_t *tmpdecodeReq = *decodeReq;
-    msgpack_object_kv* p = deserialized.via.map.ptr;
-
-    while( i < deserialized.via.map.size ) {
-        msgpack_object keyType = p->key;
-        msgpack_object ValueType = p->val;
-        keySize = keyType.via.str.size;
-        keyString = malloc( keySize + 1 );
-
-        if( keyString != NULL ) {
-            keyName = NULL;
-            keyName = getKey_MsgtypeStr( keyType, keySize, keyString );
-
-            if( keyName != NULL ) {
-                switch( ValueType.type ) {
-                    case MSGPACK_OBJECT_POSITIVE_INTEGER: {
-                        if( strcmp( keyName, WRP_MSG_TYPE.name ) == 0 ) {
-                            tmpdecodeReq->msgType = ValueType.via.i64;
-                        } else if( strcmp( keyName, WRP_STATUS.name ) == 0 ) {
-                            tmpdecodeReq->statusValue = ValueType.via.i64;
-                        } else if( strcmp( keyName, WRP_RDR.name ) == 0 ) {
-                            tmpdecodeReq->rdr = ValueType.via.i64;
-                        }
-                    }
-                    break;
-
-                    case MSGPACK_OBJECT_BOOLEAN: {
-                        if( strcmp( keyName, WRP_INCLUDE_SPANS.name ) == 0 ) {
-                            tmpdecodeReq->include_spans = ValueType.via.boolean ? true : false;
-                        }
-                    }
-                    break;
-
-                    case MSGPACK_OBJECT_STR: {
-                        StringValueSize = ValueType.via.str.size;
-                        NewStringVal = malloc( StringValueSize + 1 );
-
-                        if( NewStringVal != NULL ) {
-                            StringValue = getKey_MsgtypeStr( ValueType, StringValueSize, NewStringVal );
-
-                            if( strcmp( keyName, WRP_SOURCE.name ) == 0 ) {
-                                source = strdup( StringValue );
-
-                                if( source ) {
-                                    tmpdecodeReq->source = source;
-                                } else {
-                                    tmpdecodeReq->source = NULL;
-                                }
-                            } else if( strcmp( keyName, WRP_DEST.name ) == 0 ) {
-                                dest = strdup( StringValue );
-
-                                if( dest ) {
-                                    tmpdecodeReq->dest = dest;
-                                } else {
-                                    tmpdecodeReq->dest = NULL;
-                                }
-                            } else if( strcmp( keyName, WRP_TRANS_ID.name ) == 0 ) {
-                                transaction_uuid = strdup( StringValue );
-
-                                if( transaction_uuid ) {
-                                    tmpdecodeReq->transaction_uuid = transaction_uuid;
-                                } else {
-                                    tmpdecodeReq->transaction_uuid = NULL;
-                                }
-                            } else if( strcmp( keyName, WRP_SERVICE_NAME.name ) == 0 ) {
-                                service_name = strdup( StringValue );
-
-                                if( service_name ) {
-                                    tmpdecodeReq->service_name = service_name;
-                                } else {
-                                    tmpdecodeReq->service_name = NULL;
-                                }
-                            } else if( strcmp( keyName, WRP_URL.name ) == 0 ) {
-                                url = strdup( StringValue );
-
-                                if( url ) {
-                                    tmpdecodeReq->url = url;
-                                } else {
-                                    tmpdecodeReq->url = NULL;
-                                }
-                            } else if( strcmp( keyName, WRP_HEADERS.name ) == 0 ) {
-                                tmpdecodeReq->headers = malloc( sizeof(headers_t)
-                                                        + sizeof( char * ) * 1 );
-
-                                if( tmpdecodeReq->headers != NULL ) {
-                                    headers = strdup( StringValue );
-
-                                    if( headers ) {
-                                        tmpdecodeReq->headers->count = 1;
-                                        tmpdecodeReq->headers->headers[0] = headers;
-                                    } else {
-                                        tmpdecodeReq->headers->count = 1;
-                                        tmpdecodeReq->headers->headers[0] = NULL;
-                                    }
-                                } else {
-                                    WRP_ERROR( "Memory allocation failed\n" );
-                                }
-                            } else if( strcmp( keyName, WRP_PATH.name ) == 0 ) {
-                                path = strdup( StringValue );
-
-                                if( path ) {
-                                    tmpdecodeReq->path = path;
-                                } else {
-                                    tmpdecodeReq->path = NULL;
-                                }
-                            } else if( strcmp( keyName, WRP_CONTENT_TYPE.name ) == 0 ) {
-                                content_type = strdup( StringValue );
-
-                                if( content_type ) {
-                                    tmpdecodeReq->content_type = content_type;
-                                } else {
-                                    tmpdecodeReq->content_type = NULL;
-                                }
-                            } else if( strcmp( keyName, WRP_ACCEPT.name ) == 0 ) {
-                                accept = strdup( StringValue );
-
-                                if( accept ) {
-                                    tmpdecodeReq->accept = accept;
-                                } else {
-                                    tmpdecodeReq->accept = NULL;
-                                }
-                            } else if( strcmp( keyName, WRP_PARTNER_IDS.name ) == 0 ) {
-                                tmpdecodeReq->partner_ids = malloc( sizeof(partners_t)
-                                                            + sizeof( char * ) * 1 );
-
-                                if( tmpdecodeReq->partner_ids != NULL ) {
-                                    partnerID = strdup( StringValue );
-
-                                    if( partnerID ) {
-                                        tmpdecodeReq->partner_ids->count = 1;
-                                        tmpdecodeReq->partner_ids->partner_ids[0] = partnerID;
-                                    } else {
-                                        tmpdecodeReq->partner_ids->count = 1;
-                                        tmpdecodeReq->partner_ids->partner_ids[0] = NULL;
-                                    }
-                                } else {
-                                    WRP_ERROR( "Memory allocation failed\n" );
-                                }
-                            }
-
-                            free( NewStringVal );
-                        } else {
-                            WRP_ERROR( "Memory allocation failed\n" );
-                        }
-                    }
-                    break;
-
-                    case MSGPACK_OBJECT_BIN: {
-                        if( strcmp( keyName, WRP_PAYLOAD.name ) == 0 ) {
-                            binValueSize = ValueType.via.bin.size;
-                            payload = malloc( binValueSize + 1 );
-
-                            if( payload != NULL ) {
-                                memset( payload, 0, binValueSize + 1 );
-                                keyValue = NULL;
-                                keyValue = getKey_MsgtypeBin( ValueType, binValueSize, payload );
-
-                                if( keyValue != NULL ) {
-                                    WRP_DEBUG( "Binary payload %s\n", keyValue );
-                                }
-
-                                tmpdecodeReq->payload = keyValue;
-                                tmpdecodeReq->payload_size = binValueSize;
-                            } else {
-                                WRP_ERROR( "Memory allocation failed\n" );
-                            }
-                        }
-                    }
-                    break;
-
-                    case MSGPACK_OBJECT_ARRAY:
-
-                        if( strcmp( keyName, WRP_HEADERS.name ) == 0 ) {
-                            msgpack_object_array array = ValueType.via.array;
-                            msgpack_object *ptr = array.ptr;
-                            uint32_t cnt = 0;
-                            ptr = array.ptr;
-                            tmpdecodeReq->headers = malloc( sizeof(headers_t)
-                                                    + sizeof( char * ) * array.size );
-
-                            if( tmpdecodeReq->headers != NULL ) {
-                                tmpdecodeReq->headers->count = array.size;
-
-                                for( cnt = 0; cnt < array.size; cnt++, ptr++ ) {
-                                    tmpdecodeReq->headers->headers[cnt] = malloc( ptr->via.str.size + 1 );
-
-                                    if( tmpdecodeReq->headers->headers[cnt] != NULL ) {
-                                        memset( tmpdecodeReq->headers->headers[cnt], 0, ptr->via.str.size + 1 );
-                                        memcpy( tmpdecodeReq->headers->headers[cnt], ptr->via.str.ptr, ptr->via.str.size );
-                                        WRP_DEBUG( "tmpdecodeReq->headers[%d] %s\n", cnt, tmpdecodeReq->headers->headers[cnt] );
-                                    }
-                                }
-                            }
-                        } else if( strcmp( keyName, WRP_PARTNER_IDS.name ) == 0 ) {
-                            msgpack_object_array array = ValueType.via.array;
-                            msgpack_object *ptr = array.ptr;
-                            uint32_t cnt = 0;
-                            ptr = array.ptr;
-                            tmpdecodeReq->partner_ids = malloc( sizeof(partners_t)
-                                                        + sizeof( char * ) * array.size );
-
-                            if( tmpdecodeReq->partner_ids != NULL ) {
-                                tmpdecodeReq->partner_ids->count = array.size;
-
-                                for( cnt = 0; cnt < array.size; cnt++, ptr++ ) {
-                                    tmpdecodeReq->partner_ids->partner_ids[cnt] = malloc( ptr->via.str.size + 1 );
-
-                                    if( tmpdecodeReq->partner_ids->partner_ids[cnt] != NULL ) {
-                                        memset( tmpdecodeReq->partner_ids->partner_ids[cnt], 0, ptr->via.str.size + 1 );
-                                        memcpy( tmpdecodeReq->partner_ids->partner_ids[cnt], ptr->via.str.ptr,
-                                                ptr->via.str.size );
-                                        WRP_DEBUG( "tmpdecodeReq->partner_ids[%d] %s\n", cnt,
-                                                   tmpdecodeReq->partner_ids->partner_ids[cnt] );
-                                    }
-                                }
-                            }
-                        } else {
-                            WRP_ERROR( "Not Handled MSGPACK_OBJECT_ARRAY %s\n", keyName );
-                        }
-
-                        break;
-
-                    case MSGPACK_OBJECT_MAP:
-                        WRP_DEBUG( "Type of MAP\n" );
-                        WRP_DEBUG( "keyName is %s\n", keyName );
-                        decodeMapRequest( ValueType, decodeReq );
-                        break;
-
-                    case MSGPACK_OBJECT_NIL:
-
-                        if( strcmp( keyName, WRP_SPANS.name ) == 0 ) {
-                            WRP_DEBUG( "spans is nil\n" );
-                        }
-
-                        break;
-
-                    default:
-                        WRP_ERROR( "Unknown Data Type\n" );
-                        break;
-                }
-            }
-
-            p++;
-            i++;
-            free( keyString );
-        } else {
-            WRP_ERROR( "Memory allocation failed\n" );
-        }
-    }
-}
-
-static void decodeMapRequest( msgpack_object deserialized,
-                              struct req_res_t **decodeMapReq )
-{
-    unsigned int i = 0;
-    int n = 0, v = 0;
-    int keySize = 0;
-    char* keyString = NULL;
-    char* keyName = NULL;
-    int StringValueSize = 0;
-    char* NewStringVal, *StringValue;
-    char* mapName = NULL;
-    char *mapValue = NULL;
-    struct req_res_t *mapdecodeReq = *decodeMapReq;
-    msgpack_object_kv* p = deserialized.via.map.ptr;
-    WRP_DEBUG( "Map size is %d\n", deserialized.via.map.size );
-
-    WRP_DEBUG( "mapdecodeReq->metadata->count is %d\n", deserialized.via.map.size );
-
-    if( deserialized.via.map.size != 0 ) {
-        mapdecodeReq->metadata->count = deserialized.via.map.size;
-        mapdecodeReq->metadata->data_items = malloc( sizeof(struct data) *
-                                             ( deserialized.via.map.size ) );
-
-    }
-
-    if( mapdecodeReq->metadata->data_items != NULL ) {
-        while( i < deserialized.via.map.size ) {
-            msgpack_object keyType = p->key;
-            msgpack_object ValueType = p->val;
-            keySize = keyType.via.str.size;
-            keyString = malloc( keySize + 1 );
-
-            if( keyString != NULL ) {
-                keyName = NULL;
-                keyName = getKey_MsgtypeStr( keyType, keySize, keyString );
-
-                if( keyName != NULL ) {
-                    mapName = strdup( keyName );
-
-                    if( mapName ) {
-                        mapdecodeReq->metadata->data_items[n].name = mapName;
-                    } else {
-                        mapdecodeReq->metadata->data_items[n].name = NULL;
-                    }
-
-                    n++;
-
-                    switch( ValueType.type ) {
-                        case MSGPACK_OBJECT_POSITIVE_INTEGER: {
-                            WRP_DEBUG( "Map value is int %" PRId64 "\n", ValueType.via.i64 );
-                            sprintf( mapdecodeReq->metadata->data_items[v].value, "%" PRId64, ValueType.via.i64 );
-                            v++;
-                        }
-                        break;
-
-                        case MSGPACK_OBJECT_BOOLEAN: {
-                            WRP_DEBUG( "Map value boolean %d\n", ValueType.via.boolean ? true : false );
-                            mapdecodeReq->metadata->data_items[v].value = ValueType.via.boolean ? "true" : "false";
-                            v++;
-                        }
-                        break;
-
-                        case MSGPACK_OBJECT_STR: {
-                            StringValueSize = ValueType.via.str.size;
-                            NewStringVal = malloc( StringValueSize + 1 );
-
-                            if( NewStringVal != NULL ) {
-                                StringValue = getKey_MsgtypeStr( ValueType, StringValueSize, NewStringVal );
-                                mapValue = strdup( StringValue );
-                                mapdecodeReq->metadata->data_items[v].value = mapValue;
-                                v++;
-                                free( NewStringVal );
-                            } else {
-                                WRP_ERROR( "Memory allocation failed\n" );
-                            }
-                        }
-                        break;
-
-                        default:
-                            WRP_ERROR( "Unknown data format inside MAP\n" );
-                            break;
-                    }
-                }
-
-                p++;
-                i++;
-                free( keyString );
-            } else {
-                WRP_ERROR( "Memory allocation failed\n" );
-            }
-        }
-    } else {
-        WRP_ERROR( "Memory allocation failed\n" );
-    }
-}
-
-
-/*
-* @brief Returns the value of a given key.
-* @param[in] key key name with message type string.
-*/
-static char* getKey_MsgtypeStr( const msgpack_object key, const size_t keySize,
-                                char* keyString )
-{
-    const char* keyName = key.via.str.ptr;
-    strncpy( keyString, keyName, keySize );
-    keyString[keySize] = '\0';
-    return keyString;
-}
-
-
-/*
- * @brief Returns the value of a given key.
- * @param[in] key key name with message type binary.
- */
-static char* getKey_MsgtypeBin( const msgpack_object key, const size_t binSize,
-                                char* keyBin )
-{
-    const char* keyName = key.via.bin.ptr;
-    memcpy( keyBin, keyName, binSize );
-    return keyBin;
-}
-
-/*
-**
- *  Decode msgpack encoded data in bytes and convert to wrp struct
- *
- *  @param bytes [in]  msgpack encoded message in bytes
- *  @param length [in] length of encoded message
- *  @param msg_ptr [inout] wrp_msg_t struct initialized using bytes
- *
- *  @return the number of bytes 'consumed' by making this transformation if
- *          successful, less than 1 otherwise
- */
-
-static ssize_t __wrp_bytes_to_struct( const void *bytes, const size_t length,
-                                      wrp_msg_t **msg_ptr )
-{
-    struct req_res_t *decodeReq;
-    wrp_msg_t *msg;
-    struct req_res_t *metadata;
-
-    if( NULL == bytes ) {
-        WRP_ERROR( "bytes is NULL\n" );
-        return -1;
-    }
-
-    decodeReq = malloc( sizeof( struct req_res_t ) );
-    msg       = malloc( sizeof( wrp_msg_t ) );
-    metadata  = malloc( sizeof( data_t ) );
-
-    if( ( decodeReq != NULL ) && ( NULL != msg ) && ( NULL != metadata ) ) {
-        msgpack_zone mempool;
-        msgpack_object deserialized;
-        msgpack_unpack_return unpack_ret;
-
-        memset( msg, 0, sizeof( wrp_msg_t ) );
-        memset( decodeReq, 0, sizeof( struct req_res_t ) );
-        memset( metadata, 0, sizeof( data_t ) );
-
-        decodeReq->metadata = ( struct data_struct * ) metadata;
-        WRP_DEBUG( "unpacking encoded data\n" );
-        msgpack_zone_init( &mempool, 2048 );
-        unpack_ret = msgpack_unpack( bytes, length, NULL, &mempool, &deserialized );
-        WRP_DEBUG( "unpack_ret:%d\n", unpack_ret );
-
-        switch( unpack_ret ) {
-            case MSGPACK_UNPACK_SUCCESS:
-
-                //msgpack_object_print( stdout, deserialized );
-                //puts("");
-                if( deserialized.via.map.size != 0 ) {
-                    decodeRequest( deserialized, &decodeReq );
-                }
-
-                msgpack_zone_destroy( &mempool );
-
-                switch( decodeReq->msgType ) {
-                    case WRP_MSG_TYPE__AUTH:
-                        msg->msg_type = decodeReq->msgType;
-                        msg->u.auth.status = decodeReq->statusValue;
-                        *msg_ptr = msg;
-                        free( decodeReq->metadata->data_items );
-                        free( decodeReq->metadata );
-                        free( decodeReq );
-                        return length;
-
-                    case WRP_MSG_TYPE__REQ:
-                        msg->msg_type = decodeReq->msgType;
-                        msg->u.req.source = decodeReq->source;
-                        msg->u.req.dest = decodeReq->dest;
-                        msg->u.req.transaction_uuid = decodeReq->transaction_uuid;
-                        msg->u.req.content_type = decodeReq->content_type;
-                        msg->u.req.accept = decodeReq->accept;
-                        msg->u.req.headers = decodeReq->headers;
-                        msg->u.req.metadata = decodeReq->metadata;
-                        msg->u.req.include_spans = decodeReq->include_spans;
-                        msg->u.req.spans.spans = NULL;   /* not supported */
-                        msg->u.req.spans.count = 0;     /* not supported */
-                        msg->u.req.payload = decodeReq->payload;
-                        msg->u.req.payload_size = decodeReq->payload_size;
-                        msg->u.req.partner_ids = decodeReq->partner_ids;
-                        *msg_ptr = msg;
-                        free( decodeReq );
-                        return length;
-
-                    case WRP_MSG_TYPE__EVENT:
-                        msg->msg_type = decodeReq->msgType;
-                        msg->u.event.source = decodeReq->source;
-                        msg->u.event.dest = decodeReq->dest;
-                        msg->u.event.content_type = decodeReq->content_type;
-                        msg->u.event.metadata = decodeReq->metadata;
-                        msg->u.event.payload = decodeReq->payload;
-                        msg->u.event.payload_size = decodeReq->payload_size;
-                        msg->u.event.headers = decodeReq->headers;
-                        msg->u.event.partner_ids = decodeReq->partner_ids;
-                        *msg_ptr = msg;
-                        free( decodeReq );
-                        return length;
-
-                    case WRP_MSG_TYPE__SVC_REGISTRATION:
-                        msg->msg_type = decodeReq->msgType;
-                        msg->u.reg.service_name = decodeReq->service_name;
-                        msg->u.reg.url = decodeReq->url;
-                        *msg_ptr = msg;
-                        free( decodeReq->metadata->data_items );
-                        free( decodeReq->metadata );
-                        free( decodeReq );
-                        return length;
-
-                    case WRP_MSG_TYPE__SVC_ALIVE:
-                        msg->msg_type = decodeReq->msgType;
-                        *msg_ptr = msg;
-                        free( decodeReq->metadata );
-                        free( decodeReq );
-                        return length;
-
-                    case WRP_MSG_TYPE__CREATE:
-                    case WRP_MSG_TYPE__RETREIVE:
-                    case WRP_MSG_TYPE__UPDATE:
-                    case WRP_MSG_TYPE__DELETE:
-                        msg->msg_type = decodeReq->msgType;
-                        msg->u.crud.source = decodeReq->source;
-                        msg->u.crud.dest = decodeReq->dest;
-                        msg->u.crud.transaction_uuid = decodeReq->transaction_uuid;
-                        msg->u.crud.partner_ids = decodeReq->partner_ids;
-                        msg->u.crud.headers = decodeReq->headers;
-                        msg->u.crud.metadata = decodeReq->metadata;
-                        msg->u.crud.include_spans = decodeReq->include_spans;
-                        msg->u.crud.content_type = decodeReq->content_type;
-                        msg->u.crud.accept = decodeReq->accept;
-                        msg->u.crud.spans.spans = NULL;   /* not supported */
-                        msg->u.crud.spans.count = 0;     /* not supported */
-                        msg->u.crud.status = decodeReq->statusValue;
-                        msg->u.crud.rdr = decodeReq->rdr;
-                        msg->u.crud.payload = decodeReq->payload;
-                        msg->u.crud.payload_size = decodeReq->payload_size;
-                        msg->u.crud.path = decodeReq->path;
-                        free( decodeReq );
-                        *msg_ptr = msg;
-                        return length;
-
-                    default:
-                        free( decodeReq->metadata->data_items );
-                        break;
-                }
-
-                break;
-
-            case MSGPACK_UNPACK_EXTRA_BYTES:
-                WRP_ERROR( "MSGPACK_UNPACK_EXTRA_BYTES\n" );
-                break;
-
-            case MSGPACK_UNPACK_CONTINUE:
-                WRP_ERROR( "MSGPACK_UNPACK_CONTINUE\n" );
-                break;
-
-            case MSGPACK_UNPACK_PARSE_ERROR:
-                WRP_ERROR( "MSGPACK_UNPACK_PARSE_ERROR\n" );
-                break;
-
-            case MSGPACK_UNPACK_NOMEM_ERROR:
-                WRP_ERROR( "MSGPACK_UNPACK_NOMEM_ERROR\n" );
-                break;
-
-            default:
-                break;
-        }
-
-        msgpack_zone_destroy( &mempool );
-    } else {
-        WRP_ERROR( "Memory allocation failed\n" );
-    }
-
-    if( NULL != decodeReq ) {
-        free( decodeReq );
-    }
-
-    if( NULL != msg ) {
-        free( msg );
-    }
-
-    if( NULL != metadata ) {
-        free( metadata );
-    }
-
-    return -1;
-}
-
-/*
-**
- *  Decode base64 encoded msgpack encoded data in bytes and convert to wrp struct
- *
- *  @param bytes [in]  base64 encoded msgpack encoded message in bytes
- *  @param length [in] length of base64 encoded message
- *  @param msg_ptr [inout] wrp_msg_t struct initialized
- *
- *  @return the number of bytes 'consumed' by making this transformation if
- *          successful, less than 1 otherwise
- */
-static ssize_t __wrp_base64_to_struct( const void *base64_data, const size_t base64_size,
-                                       wrp_msg_t **msg_ptr )
-{
-    ssize_t rv = -1;
-    size_t length = 0;
-    uint8_t *bytes = NULL;
-
-    bytes = b64_decode_with_alloc( (const uint8_t*) base64_data, base64_size, &length );
-    if( bytes ) {
-        rv = __wrp_bytes_to_struct( bytes, length, msg_ptr );
-        free( bytes );
-    }
-
-    return rv;
-}
 
 
 /**
@@ -1984,7 +484,6 @@ static ssize_t __wrp_base64_to_struct( const void *base64_data, const size_t bas
  * @param[in] encodedBuffer msgpack object
  * @param[out] return 0 in success or less than 1 in failure case
  */
-
 static int alterMap( char * buf )
 {
     //Extract 1st byte from binary stream which holds type and map size
@@ -2008,6 +507,7 @@ static int alterMap( char * buf )
     return 0;
 }
 
+
 /**
  * @brief appendEncodedData function to append two encoded buffer and change MAP size accordingly.
  *
@@ -2021,7 +521,6 @@ static int alterMap( char * buf )
  * @param[out] appendData final encoded buffer after append
  * @return  appended total buffer size or less than 1 in failure case
  */
-
 size_t appendEncodedData( void **appendData, void *encodedBuffer, size_t encodedSize,
                           void *metadataPack, size_t metadataSize )
 {
