+++ tests/CMakeLists.txt
@@ -10,8 +10,11 @@ set (CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -W  -g -fprofile-arcs -ftest-cove
 set (CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} -fprofile-arcs -ftest-coverage -O0 -finstrument-functions")
 
 add_test(NAME Simple COMMAND ${MEMORY_CHECK} ./simple)
-add_executable(simple simple.c ../src/wrp-c.c
-                               ../src/utils.c)
+add_executable(simple simple.c ../src/b64.c
+                               ../src/msgpack.c
+                               ../src/string.c
+                               ../src/utils.c
+                               ../src/wrp-c.c)
 target_link_libraries (simple ${TROWER_LIBRARIES}
                               ${CIMPLOG_LIBRARIES}
                               ${MSGPACK_LIBRARIES}
